refactor(core): split the x402 listing route tests by concern
The suite reached 839 lines — the round-3 log recorded it at 715 with
"should be split by concern before it grows", and this round grew it.

The boundary is the one the file already had implicitly:
`get.query.test.ts` covers how the route ASKS Postgres for its page
(pagination bounds, cursor and its id tiebreak, relation ordering, the
RepeatableRead snapshot, column narrowing); `get.test.ts` keeps
authorization, the fail-closed per-agent gates, and drop logging.

Row builders and the app harness move to `get.fixtures.ts` — not a
`*.test.ts` name, so vitest's include does not collect it. Only the
`vi.hoisted` mock objects stay duplicated per file, which `vi.mock`
factory scoping requires. No test was added, removed, or reworded: 25
before, 16 + 9 after.

Moving `COWORKER_AGENT_CONTEXT` across a module boundary surfaced a
latent typing weakness: annotated with the whole `AuthenticationContext`
union it only compiled because an in-file `const` narrows back to its
initializer's member, so the delegated-coworker spread of `context`
failed once imported. It is now typed as `CoworkerAuthenticationContext`,
which is what it always was.
